phase3/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "AI Agent Service for Todo List App"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai-agents>=0.3.2",
    "python-mcp>=1.0.0",
    "httpx>=0.27.2",
    "pydantic>=2.10",
    "pydantic-settings>=2.6.1",
    "openai-chatkit>=1.4.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "todo_agent", "ai_mcp_server", "main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "models*"]

[tool.uv.workspace]
members = [
    "ai-service",
]
